// src/types/editor2-types.ts
export interface BlockComponentProps {
  id?: string;
  children?: React.ReactNode;
  className?: string;
  responsiveStyles?: Record<string, any>;
  styles?: Record<string, any>;
  style?: React.CSSProperties;
}